% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dp_get_dataset}
\alias{dp_get_dataset}
\title{Download a dataset given the name and optionally the url}
\usage{
dp_get_dataset(set_name, set_url = "")
}
\arguments{
\item{data_set_name}{a string}

\item{data_set_url}{a string, optional}
}
\value{
A list with a csv file and a json of extra info
}
\description{
Given a dataset name, and optionally the url in the dati.piemonte portal it downloads the raw data.
It is left to the user what to do with those.
}
