% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dp_get_page_url}
\alias{dp_get_page_url}
\title{Get the url of a dataset given the name}
\usage{
dp_get_page_url(datasetname)
}
\value{
A string
}
\description{
This is an internal functionality.
Given the name of a dataset it downloads the catalogue and check for the corresponding url.
}
\examples{

dp_get_page_url("datasetname from dp_list_datasets()")
}
