% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dp_list_datasets}
\alias{dp_list_datasets}
\title{Get the availabe datasets on dati.piemonte}
\usage{
dp_list_datasets()
}
\value{
A tibble
}
\description{
This functions retrieves the file that list the open data on the portal and reads it.
Currently only YUCCA hosted datasets are supported, if you want to add more, do a pull request.
You can use the data_url for further calls, of just the name, is up to you.
}
\examples{
dp_list_datasets()
}
