% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dp_process_page_url}
\alias{dp_process_page_url}
\title{Make the download and documentation links from a page url}
\usage{
dp_process_page_url(page_url)
}
\arguments{
\item{page_url}{the url of the dataset page on the datipiemonte.it portal}
}
\value{
A list with 3 urls
}
\description{
This is an internal functionality.
Given a page url it creates urls to download data and metadata.
Currently only supports the YUCCA hosted datasets.
}
\examples{
if(FALSE){dp_process_page_url(page_url)}
}
