% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dp_load_dataset}
\alias{dp_load_dataset}
\title{Cleans and load a dataset object returned from dp_get_dataset
To be used if the user wants to avoid manual inspection
May fail}
\usage{
dp_load_dataset(raw_dataset_object)
}
\arguments{
\item{raw_dataset_object}{the output of dp_get_dataset}
}
\value{
A tibble
}
\description{
This is a utility for lazy people.
It takes as input a list outputted by dp_get_dataset and uses the metadata to define the names.
Columns encoding is guesses.
The separator is guessed to be a semicolon, can be wrong.
If you want to contribute in making this function better, do a pull request.
}
\examples{
dataset_name \%>\% dp_get_dataset \%>\% dp_load_dataset
}
